refactor(d9): replace new_rope with prev since the former isn't needed?
It seems like it wasn't actually necessary to create new_rope, since we iterate over &mut and moving the knots moves them in the original rope anyway. The only thing we really needed was the previous knot, to measure the distance from it. But we can't take a reference to it because ownership (or so it seems?) - so just store it